Updated train_actor.py

+++ mt/erac/train_actor.py
@@ -174,7 +174,7 @@ def evaluate(iterator):
 
         # compute masked nll
         tgt_flat = tgt[1:].view(-1, 1)
-        masked_nll = -log_prob.view(-1, ntoken_tgt).gather(1, tgt_flat).masked_select(tgt_flat.ne(tgt_pad_idx))
+        masked_nll = -log_prob.view(-1, ntoken_tgt).gather(1, tgt_flat.to(torch.int64)).masked_select(tgt_flat.ne(tgt_pad_idx))
 
         # accumulate nll
         sum_nll += masked_nll.data.sum()
